+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Ongoing
 
+- Improve node removal: remove and reset the node as executed by Source, and remove the cache-file.
 - PR [342](https://github.com/plugwise/python-plugwise-usb/pull/342): Improve node_type chaching.
-- Improve node removal: reset the node before removing and remove the cache-file too.
 
 ## 0.46.0 - 2025-09-12